requirements.txt
PyQt5
requests
psutil
matplotlib
python-dotenv
